Refactors auth feature structure for improved clarity
Organizes authentication-related files into more consistent directories and naming schemes.
Unifies API helpers, form schemas, and hooks naming conventions for better maintainability.
Improves codebase scalability and reduces confusion for future development.
