+++ src/app/(routes)/dashboard/page.tsx
@@ -20,8 +20,8 @@ import {
   IconButton,
 } from "@mui/material";
 import { FiLogOut } from "react-icons/fi"; // Using a react-icon for the logout button
-import { useCurrentUserQuery } from "@/src/features/auth/use_current_user_query";
-import { useLogoutMutation } from "@/src/features/auth/use_logout_mutation";
+import { useCurrentUserQuery } from "@/src/hooks/use_current_user_query";
+import { useLogoutMutation } from "@/src/features/auth/hooks/use_logout_mutation";
 
 // This is a placeholder type. In a real app, you'd have a proper User type.
 type User = {
