+++ src/app/(routes)/login/page.tsx
@@ -3,10 +3,10 @@
 import { useRouter } from "next/navigation";
 import { APP_ROUTES } from "@/src/constants/paths";
 import { Toaster, toast } from "sonner";
-import { useLoginMutation } from "../../../features/auth/use_login_mutation";
-import LoginForm from "../../../features/login/components/login_form";
-import type { LoginFormData } from "../../../features/login/form_schema";
+import LoginForm from "../../../features/auth/components/login_form";
+import type { LoginFormData } from "../../../features/auth/constants/login_form_schema";
 import { Box } from "@mui/material";
+import { useLoginMutation } from "@/src/features/auth/hooks/use_login_mutation";
 
 export default function LoginPage() {
   const router = useRouter();
