+++ src/app/(routes)/signup/page.tsx
@@ -3,9 +3,9 @@
 import { Box } from "@mui/material";
 import { useRouter } from "next/navigation";
 import { APP_ROUTES } from "@/src/constants/paths";
-import SignupForm from "../../../features/signup/signup_form";
-import type { SignUpFormData } from "../../../features/signup/schema";
-import { useSignUpMutation } from "../../../features/auth/use_sign_up_mutation";
+import SignupForm from "../../../features/auth/components/signup_form";
+import type { SignUpFormData } from "../../../features/auth/constants/signup_form_schema";
+import { useSignUpMutation } from "../../../features/auth/hooks/use_sign_up_mutation";
 
 export default function SignUpPage() {
   const router = useRouter();
