+++ src/app/api/login/route.ts
@@ -1,10 +1,13 @@
-import { createSession, verifyUser } from "@/src/features/auth/libs/auth";
+import {
+  createSession,
+  verifyUser,
+} from "@/src/features/auth/libs/session_manager";
 import { NextRequest, NextResponse } from "next/server";
 import z from "zod";
 import {
   createSuccessResponse,
   createErrorResponse,
-} from "@/src/lib/api_helpers";
+} from "@/src/lib/api_response_helpers";
 
 const SESSION_DURATION_SECONDS = 60 * 60 * 24; // 1 day
